app/server/pyproject.toml
[project]
name = "edge-server"
version = "0.1.0"
description = "Learning project for a2ui"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "a2ui>=0.1.0",
    "fastapi>=0.110.0",
    "jsonschema>=4.26.0",
    "langchain>=1.2.3",
    "langchain-core>=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-oci>=0.2.1",
    "mcp>=1.25.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "oci-openai>=1.0.0",
    "playwright>=1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv.sources]
a2ui = { path = "../../a2a_agents/python/a2ui_extension", editable = true }
